requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
scikit-learn==1.3.2
PyPDF2==3.0.1
Pillow==10.1.0
openpyxl==3.1.2
pdf2image==1.16.3
werkzeug==3.0.1